client/requirements.txt
<<<<<<< HEAD
boto3

=======
flask==1.1.2
flask-restful==0.3.8
flask-sqlalchemy==2.4.1
flask-jsonrpc==0.3.1
requests==2.23.0
werkzeug==1.0.1
flask_cors==3.0.8
cryptography==2.9
>>>>>>> a25bdf8e200854e90a148f5d6c1bead32df85226
